+++ node/tests/TestUtilities.ts
@@ -875,6 +875,11 @@ export async function transactionTest(
         'xtrim(key9, { method: "minid", threshold: "0-2", exact: true }',
         1,
     ]);
+    baseTransaction.customCommand(["XGROUP", "CREATE", key9, "group", "0-0"]);
+    responseData.push(['xgroupCreate(key9, group, "0-0")', "OK"]);
+    baseTransaction.xinfoConsumers(key9, "group");
+    responseData.push(['xinfoConsumers(key9, "group")', []]);
+
     baseTransaction.xdel(key9, ["0-3", "0-5"]);
     responseData.push(["xdel(key9, [['0-3', '0-5']])", 1]);
     baseTransaction.rename(key9, key10);